requirements.txt
asgiref==3.8.1
Django==5.0.4
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-log-request-id==2.1.0
django-request-id==1.0.0
djangorestframework==3.15.1
Markdown==3.6
ruff==0.4.3
sqlparse==0.5.0
typing_extensions==4.11.0
